fix dpsf

+++ src/vggt/heads/camera_head.py
@@ -93,10 +93,10 @@ class CameraHead(nn.Module):
             list: A list of predicted camera encodings (post-activation) from each iteration.
         """
         # Use tokens from the last block for camera prediction.
-        tokens = aggregated_tokens_list[-1]
+        tokens = aggregated_tokens_list[-1].float()
 
         # Extract the camera tokens
-        pose_tokens = tokens[:, :, 0]
+        pose_tokens = tokens[:, :, 0].float()
         pose_tokens = self.token_norm(pose_tokens)
 
         pred_pose_enc_list = self.trunk_fn(pose_tokens, num_iterations)
